src/requirements.txt
Flask
python-dotenv
azure-identity==1.21.0
azure-ai-projects==1.0.0b8
